+++ tests/test_client.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from httplib import CONFLICT, NOT_FOUND, PRECONDITION_FAILED, UNAUTHORIZED
 import json
 from tests import FileObjMatcher, UTC, mocked_response
-import unittest
+import unittest2 as unittest
 
 from flexmock import flexmock
 import requests
@@ -204,14 +204,13 @@ class TestClient(unittest.TestCase):
         result = client.get_user_list(limit=123, offset=456)
         self.assertEqual({'a': 'b'}, result)
 
-
     def test_get_folder(self):
         client = self.make_client("get", 'folders/666', params={'limit': 123, 'offset': 456}, result={'a': 'b'})
         result = client.get_folder(folder_id=666, limit=123, offset=456)
         self.assertEqual({'a': 'b'}, result)
 
-        client = self.make_client("get", 'folders/666', params={'limit': 123, 'offset': 456, 'fields': ['hello']}, result={'a': 'b'})
-        result = client.get_folder(folder_id=666, limit=123, offset=456, fields=['hello'])
+        client = self.make_client("get", 'folders/666', params={'limit': 123, 'offset': 456, 'fields': 'hello,goodbye'}, result={'a': 'b'})
+        result = client.get_folder(folder_id=666, limit=123, offset=456, fields=['hello', 'goodbye'])
         self.assertEqual({'a': 'b'}, result)
 
     def test_get_folder_content(self):
@@ -219,7 +218,7 @@ class TestClient(unittest.TestCase):
         result = client.get_folder_content(folder_id=666, limit=123, offset=456)
         self.assertEqual({'a': 'b'}, result)
 
-        client = self.make_client("get", 'folders/666/items', params={'limit': 123, 'offset': 456, 'fields': ['hello']}, result={'a': 'b'})
+        client = self.make_client("get", 'folders/666/items', params={'limit': 123, 'offset': 456, 'fields': 'hello'}, result={'a': 'b'})
         result = client.get_folder_content(folder_id=666, limit=123, offset=456, fields=['hello'])
         self.assertEqual({'a': 'b'}, result)
 
@@ -229,16 +228,43 @@ class TestClient(unittest.TestCase):
         (flexmock(client)
             .should_receive('get_folder_content')
             .with_args(666, limit=1000)
-            .and_return({'entries': range(10)})
+            .and_return({'entries': range(10), 'total_count': 10})
+            .once())
+
+        self.assertSequenceEqual(list(client.get_folder_iterator(666)), range(10))
+
+    def test_get_folder_iterator_boundary_1(self):
+        # setup a regular client without expecting the usual calls
+        client = BoxClient('my_token')
+        (flexmock(client)
+            .should_receive('get_folder_content')
+            .with_args(666, limit=1000)
+            .and_return({'entries': range(1000), 'total_count': 1001})
             .once())
 
         (flexmock(client)
             .should_receive('get_folder_content')
             .with_args(666, limit=1000, offset=1000)
-            .and_return({'entries': None})
+            .and_return({'entries': [1000], 'total_count': 1001})
             .once())
 
-        self.assertListEqual(list(client.get_folder_iterator(666)), list(range(10)))
+        self.assertSequenceEqual(list(client.get_folder_iterator(666)), range(1001))
+
+    def test_get_folder_iterator_boundary_2(self):
+        # setup a regular client without expecting the usual calls
+        client = BoxClient('my_token')
+        (flexmock(client)
+            .should_receive('get_folder_content')
+            .with_args(666, limit=1000)
+            .and_return({'entries': range(1000), 'total_count': 1000})
+            .once())
+
+        (flexmock(client)
+            .should_receive('get_folder_content')
+            .with_args(666, limit=1000, offset=1000)
+            .never())
+
+        self.assertSequenceEqual(list(client.get_folder_iterator(666)), range(1000))
 
     def test_get_folder_iterator_zero_content(self):
         # setup a regular client without expecting the usual calls
@@ -250,6 +276,19 @@ class TestClient(unittest.TestCase):
             .once())
 
         self.assertListEqual(list(client.get_folder_iterator(666)), [])
+
+    def test_get_folder_collaborations(self):
+        client = self.make_client("get", 'folders/123/collaborations', result={'a': 'b'})
+        self.assertEqual({'a': 'b'}, client.get_folder_collaborations(123))
+
+    def test_copy_folder(self):
+        client = self.make_client("post", 'folders/123/copy', data={'parent': {'id': '666'}}, result={'id': '1'})
+        result = client.copy_folder(123, 666)
+        self.assertEqual({'id': '1'}, result)
+
+        client = self.make_client("post", 'folders/123/copy', data={'parent': {'id': '666'}, 'name': 'goatse.cx'}, result={'id': '1'})
+        result = client.copy_folder(123, 666, 'goatse.cx')
+        self.assertEqual({'id': '1'}, result)
 
     def test_create_folder_no_parent(self):
         expected_dict = {
@@ -341,32 +380,379 @@ class TestClient(unittest.TestCase):
                        'https://api.box.com/2.0/files/123/thumbnail.png',
                        params={},
                        data=None,
-                       headers=client.default_headers)
+                       headers=client.default_headers,
+                       stream=True)
             .and_return(mocked_response(status_code=202, headers={"Location": "http://box.com", "Retry-After": "5"}))
             .once())
 
         thumbnail = client.get_thumbnail(123)
         self.assertIsNone(thumbnail)
 
-        # Delayed within allowed wait time
+    def test_file_get_comments(self):
+        client = BoxClient("my_token")
+
+        response = { "total_count": 0, "entries": [] }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                       'https://api.box.com/2.0/files/123/comments',
+                       params=None,
+                       data=None,
+                       headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        comments = client.get_file_comments(123)
+        self.assertEquals(comments, response)
+
+    def test_get_comment_information(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "comment",
+                    "id": 123
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                 "https://api.box.com/2.0/comments/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        comment = client.get_comment_information(123)
+        self.assertEquals(comment, response)
+
+    def test_add_comment_to_file(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "comment",
+                    "id": 123,
+                    "item": {"id": 123,
+                             "type": "file"},
+                    "message": "test"
+        }
+
+        expected_data={"item": {"type": "file",
+                                "id": 123},
+                       "message": "test"
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("post",
+                 "https://api.box.com/2.0/comments",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        comment = client.add_comment(123, "file", "test")
+        self.assertEquals(comment, response)
+
+    def test_add_comment_to_comment(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "comment",
+                    "id": 123,
+                    "item": {"id": 123,
+                             "type": "comment"},
+                    "message": "test"
+        }
+
+        expected_data={"item": {"type": "comment",
+                                "id": 123},
+                       "message": "test"
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("post",
+                 "https://api.box.com/2.0/comments",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        comment = client.add_comment(123, "comment", "test")
+        self.assertEquals(comment, response)
+
+    def test_change_comment(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "comment",
+                    "id": 123,
+                    "message": "new_message"
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("put",
+                 "https://api.box.com/2.0/comments/123",
+                 params=None,
+                 data=json.dumps({"message": "new_message"}),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        modified = client.change_comment(123, "new_message")
+        self.assertEquals(modified, response)
+
+    def test_delete_comment(self):
+        client = BoxClient("my_token")
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("delete",
+                 "https://api.box.com/2.0/comments/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(status_code=204)))
+
+        self.assertIsNone(client.delete_comment(123))
+
+    def test_file_get_tasks(self):
+        client = BoxClient("my_token")
+
+        response = { "total_count": 0, "entries": [] }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                       'https://api.box.com/2.0/files/123/tasks',
+                       params=None,
+                       data=None,
+                       headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        tasks = client.get_file_tasks(123)
+        self.assertEquals(tasks, response)
+
+    def test_get_task_information(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "task",
+                    "id": 123
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                 "https://api.box.com/2.0/tasks/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        task = client.get_task_information(123)
+        self.assertEquals(task, response)
+
+    def test_add_task(self):
+        client = BoxClient("my_token")
+        due_at = datetime.now()
+
+        expected_data = {"item": {"type": "file",
+                                  "id": 123},
+                         "action": "review",
+                         "due_at": str(due_at),
+                         "message": "test"
+        }
+
+        response = {"type": "task",
+                    "id": 123,
+                    "action": "review",
+                    "message": "test",
+                    "due_at": str(due_at)
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("post",
+                 "https://api.box.com/2.0/tasks",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        task = client.add_task(123, due_at, message="test")
+        self.assertEquals(task, response)
+
+    def test_change_task(self):
+        client = BoxClient("my_token")
+        due_at = datetime.now()
+
+        expected_data = {"action": "review",
+                         "due_at": str(due_at),
+                         "message": "changed"
+        }
+
+        response = {"type": "task",
+                    "id": 123,
+                    "action": "review",
+                    "message": "changed",
+                    "due_at": str(due_at)
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("put",
+                 "https://api.box.com/2.0/tasks/123",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        changed = client.change_task(123, due_at, message="changed")
+        self.assertEquals(changed, response)
+
+    def test_delete_task(self):
+        client = BoxClient("my_token")
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("delete",
+                 "https://api.box.com/2.0/tasks/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(status_code=204)))
+
+        self.assertIsNone(client.delete_task(123))
+
+    def test_get_task_assignments(self):
+        client = BoxClient("my_token")
+
+        response = {"total_count": 0,
+                    "entries": []
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                 "https://api.box.com/2.0/tasks/123/assignments",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        assignments = client.get_task_assignments(123)
+        self.assertEquals(assignments, response)
+
+    def test_get_assignment_information(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "task_assignment",
+                    "id": 123
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("get",
+                 "https://api.box.com/2.0/task_assignments/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        assignment = client.get_assignment(123)
+        self.assertEquals(assignment, response)
+
+    def test_add_assignment(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "task_assignment",
+                    "id": 123,
+                    "assigned_to": {"type": "user",
+                                    "id": 123,
+                                    "login": "test@test.com"},
+                    "item": {"type": "task",
+                             "id": 123}
+        }
+
+        expected_data = {"task": {"id": 123,
+                                  "type": "task"},
+                         "assign_to": {"id": 123,
+                                       "login": "test@test.com"}
+        }
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("post",
+                 "https://api.box.com/2.0/task_assignments",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        assignment = client.assign_task(123, user_id=123, login="test@test.com")
+        self.assertEquals(assignment, response)
+
+    def test_update_assignment(self):
+        client = BoxClient("my_token")
+
+        response = {"type": "task_assignment",
+                    "id": 123,
+                    "message": "All good !!!",
+                    "resolution_state": "completed",
+                    "assigned_to": {"type": "user",
+                                    "id": 123,
+                                    "login": "test@test.com"},
+                    "item": {"type": "task",
+                             "id": 123}
+        }
+
+        expected_data = {"resolution_state": "completed",
+                         "message": "All good !!!"}
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("put",
+                 "https://api.box.com/2.0/task_assignments/123",
+                 params=None,
+                 data=json.dumps(expected_data),
+                 headers=client.default_headers)
+        .and_return(mocked_response(response)))
+
+        changed = client.update_assignment(123, "completed", "All good !!!")
+        self.assertEquals(changed, response)
+
+    def test_delete_assignment(self):
+        client = BoxClient("my_token")
+
+        (flexmock(requests)
+            .should_receive('request')
+            .with_args("delete",
+                 "https://api.box.com/2.0/task_assignments/123",
+                 params=None,
+                 data=None,
+                 headers=client.default_headers)
+        .and_return(mocked_response(status_code=204)))
+
+        self.assertIsNone(client.delete_assignment(123))
+
+    def test_get_client_with_retry(self):
+        client = BoxClient("my_token")
+
+
+
         (flexmock(requests)
             .should_receive('request')
             .with_args("get",
                        'https://api.box.com/2.0/files/123/thumbnail.png',
                        params={},
                        data=None,
-                       headers=client.default_headers)
-            .and_return(mocked_response(status_code=202, headers={"Location": "http://box.com/url_to_thumbnail", "Retry-After": "1"}))
-            .once())
-
-        (flexmock(requests)
-            .should_receive('get')
-            .with_args('http://box.com/url_to_thumbnail', headers=client.default_headers)
-            .and_return(mocked_response(StringIO("Thumbnail contents")))
-            .once())
+                       headers=client.default_headers,
+                       stream=True)
+            .and_return(mocked_response(status_code=202, headers={"Location": "http://box.com/url_to_thumbnail", "Retry-After": "1"}),
+                        mocked_response(StringIO("Thumbnail contents")))
+            .one_by_one())
 
         thumbnail = client.get_thumbnail(123, max_wait=1)
         self.assertEqual('Thumbnail contents', thumbnail.read())
+
+    def test_get_thumbnail_with_params(self):
+        client = BoxClient("my_token")
 
         # Not available
         (flexmock(requests)
@@ -375,7 +761,8 @@ class TestClient(unittest.TestCase):
                        'https://api.box.com/2.0/files/123/thumbnail.png',
                        params={},
                        data=None,
-                       headers=client.default_headers)
+                       headers=client.default_headers,
+                       stream=True)
             .and_return(mocked_response(status_code=302))
             .once())
 
@@ -389,7 +776,8 @@ class TestClient(unittest.TestCase):
                        'https://api.box.com/2.0/files/123/thumbnail.png',
                        params={},
                        data=None,
-                       headers=client.default_headers)
+                       headers=client.default_headers,
+                       stream=True)
             .and_return(mocked_response(StringIO("Thumbnail contents")))
             .once())
 
@@ -406,7 +794,8 @@ class TestClient(unittest.TestCase):
                                "min_width": 3,
                                "max_width": 4},
                        data=None,
-                       headers=client.default_headers)
+                       headers=client.default_headers,
+                       stream=True)
             .and_return(mocked_response(StringIO("Thumbnail contents")))
             .once())
 
@@ -774,6 +1163,56 @@ class TestClient(unittest.TestCase):
         client = self.make_client("get", "search", params={'query': "foobar", 'limit': 123, 'offset': 456}, result=expected_result)
         result = client.search("foobar", limit=123, offset=456)
         self.assertEqual(result, expected_result)
+
+    def test_get_collaboration(self):
+        client = self.make_client("get", 'collaborations/123', result={'a': 'b'})
+        self.assertEqual({'a': 'b'}, client.get_collaboration(123))
+
+    def test_create_collaboration_by_user_id(self):
+        params = {
+            'notify': False,
+        }
+        data = {
+            'item': {'id': 123, 'type': 'folder'},
+            'accessible_by': {'id': 123, 'type': 'user'},
+            'role': 'viewer',
+        }
+        expected_result = {'entries': None}
+        client = self.make_client("post", "collaborations", params, data=data, result=expected_result)
+        result = client.create_collaboration_by_user_id(123, 123)
+        self.assertEqual(result, expected_result)
+
+    def test_create_collaboration_by_login(self):
+        params = {
+            'notify': False,
+        }
+        data = {
+            'item': {'id': 123, 'type': 'folder'},
+            'accessible_by': {'login': 'sean@box.com', 'type': 'user'},
+            'role': 'viewer',
+        }
+        expected_result = {'entries': None}
+        client = self.make_client("post", "collaborations", params, data=data, result=expected_result)
+        result = client.create_collaboration_by_login(123, 'sean@box.com')
+        self.assertEqual(result, expected_result)
+
+    def test_edit_collaboration(self):
+        data = {
+            'role': 'viewer',
+        }
+        expected_result = {'entries': None}
+        client = self.make_client('put', 'collaborations/123', data=data, result=expected_result)
+        result = client.edit_collaboration(123)
+        self.assertEqual(result, expected_result)
+
+    def test_delete_collaboration(self):
+        client = self.make_client("delete", 'collaborations/123')
+        result = client.delete_collaboration(123)
+        self.assertIsNone(result)
+
+        client = self.make_client("delete", 'collaborations/123', headers={'If-Match': 'deadbeef'})
+        result = client.delete_collaboration(123, etag='deadbeef')
+        self.assertIsNone(result)
 
 
 if __name__ == '__main__':